add serialization to server/client
